Extract config to .env (#49)
