+++ web-previewer/.env.example
@@ -1,0 +1,4 @@
+GENESIS_CONTRACT='0x25ed58c027921E14D86380eA2646E3a1B5C55A8b'
+PIXEL_AVATAR_CONTRACT='0x8F0EEfABBF5164453D08081c063c1408d08f8c7f'
+NETWORK='rinkeby'
+INFURA_ID=
