+++ web-previewer/src/config.js
@@ -1,0 +1,4 @@
+export const GENESIS_CONTRACT = process.env.VUE_APP_GENESIS_CONTRACT
+export const PIXEL_AVATAR_CONTRACT = process.env.VUE_APP_PIXEL_AVATAR_CONTRACT
+export const NETWORK = process.env.VUE_APP_NETWORK
+export const INFURA_ID = process.env.VUE_APP_INFURA_ID
