--- web-previewer/src/constants/adresses.js
@@ -1,11 +1,0 @@
-//Developer Daos Genesis contract
-export const GENESIS_CONTRACT = '0x25ed58c027921E14D86380eA2646E3a1B5C55A8b'
-
-export const PIXEL_AVATAR_CONTRACT =
-    '0x8F0EEfABBF5164453D08081c063c1408d08f8c7f'
-
-export const NETWORK = 'rinkeby'
-
-//TODO create INFURA_ID with D_D account
-//Alex245's INFURA_ID
-export const INFURA_ID = 'ed20da14e0ed4fe4989da6cc2de4c99e'
